fix: remove duplicated interface
