+++ packages/types/src/lib/node-interfaces/node-interfaces.ts
@@ -2,7 +2,11 @@
  * This file serves as a central location for all Lit node endpoints and their request/response interfaces & types.
  */
 
-import { AuthMethod, AuthSig } from '../interfaces';
+import {
+  AuthMethod,
+  AuthSig,
+  MultipleAccessControlConditions,
+} from '../interfaces';
 import {
   AccessControlConditions,
   EvmContractConditions,
@@ -31,16 +35,16 @@ export interface JsonPKPClaimKeyRequest {
 //     #[serde(default = "default_epoch")]
 //     pub epoch: u64,
 // }
-export interface SigningAccessControlConditionRequest {
-  accessControlConditions?: AccessControlConditions[];
-  evmContractConditions?: EvmContractConditions[];
-  solRpcConditions?: SolRpcConditions[];
-  unifiedAccessControlConditions?: UnifiedAccessControlConditions[];
-  chain?: string | null;
-  authSig: AuthSig;
-  iat: number;
-  exp: number;
-  epoch: number;
+export interface SigningAccessControlConditionRequest
+  extends MultipleAccessControlConditions {
+  // The chain name of the chain that you are querying.  See ALL_LIT_CHAINS for currently supported chains.
+  chain?: string;
+
+  // The authentication signature that proves that the user owns the crypto wallet address that meets the access control conditions
+  authSig?: AuthSig;
+
+  iat?: number;
+  exp?: number;
 }
 
 // pub struct EncryptionSignRequest {
